cmidi2_ump_get_sysex8_data(): fix packet iteration and reported data size.
The "number of packets" is incorrect and not useful when the length of each
packet varies (4 bytes, 8 bytes, 16 bytes). Return the number of
`cmidi2_ump`s for API usability.

It was also iterating the packets in wrong way; continuous sysex8/MDS
packets should be iterated per 16 bytes. It was thus resulting incomplete
iteration.

The resulting data length was also wrong, as it contained that of streamId.
It is precise as in UMP, not as in the binary data size.